Contrive Dbugingrad clause as a new statute
Why? Dbugingrad was created but exists only as a PDF of a particular treaty that isn't actively tracked. But given the peculiarities of the Dbugingrad, it cannot be made part of the Red Rocks Pact, nor the Statehood of Icarus agreement. As such, the only logical outcome is contriving it into its own statute.